requirements.txt
influxdb-client
python-dotenv
psutil